requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
jinja2==3.1.2
trakt==3.4.0
tmdbsimple==2.9.1
httpx==0.24.1
httpcore==0.17.3 